agent accuracy is increased

+++ backend/internal/handlers/server_handler.go
@@ -42,8 +42,8 @@ type ServerRegisterRequest struct {
 	Architecture    string `json:"architecture"`
 	MACAddress      string `json:"mac_address"`
 	CPUModel        string `json:"cpu_model"`
-	TotalMemoryGB   uint64 `json:"total_memory_gb"`
-	TotalDiskGB     uint64 `json:"total_disk_gb"`
+	TotalMemoryGB   float64 `json:"total_memory_gb"`
+	TotalDiskGB     float64 `json:"total_disk_gb"`
 	GPU             string `json:"gpu"`
 	Virtualization  string `json:"virtualization"`
 	CloudProvider   string `json:"cloud_provider"`
