+++ backend/internal/services/machine_service.go
@@ -26,8 +26,8 @@ type RegisterMachineInput struct {
 	Architecture   string
 	MACAddress     string
 	CPUModel       string
-	TotalMemoryGB  uint64
-	TotalDiskGB    uint64
+	TotalMemoryGB  float64
+	TotalDiskGB    float64
 	GPU            string
 	Virtualization string
 	CloudProvider  string
